+++ jax/scipy/interpolate/__init__.py
@@ -14,5 +14,13 @@
 
 # Already deprecate namespaces that will be removed in SciPy v2.0.0
 
+
+from jax._src.scipy.interpolate.cubic import (
+    Akima1DInterpolator as Akima1DInterpolator,
+    CubicHermiteSpline as CubicHermiteSpline)
+
+from jax._src.scipy.interpolate.ppoly import (
+    PPoly as PPoly)
+
 from jax._src.third_party.scipy.interpolate import (
     RegularGridInterpolator as RegularGridInterpolator)